refactor: Extract awaitJobCompletion to shared ExecutorTestUtil
