+++ app/src/test/java/ai/brokk/executor/SearchModeSearchAgentTest.java
@@ -1,5 +1,6 @@
 package ai.brokk.executor;
 
+import static ai.brokk.testutil.ExecutorTestUtil.awaitJobCompletion;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -110,7 +111,7 @@ class SearchModeSearchAgentTest {
         var jobId = createJobWithSpec(jobSpec, "search-test-read-only");
 
         // Wait for job to complete
-        awaitJobCompletion(jobId, Duration.ofSeconds(30));
+        awaitJobCompletion(baseUrl, jobId, authToken, Duration.ofSeconds(30));
 
         // Poll for events to ensure job has processed
         var eventsUrl = URI.create(baseUrl + "/v1/jobs/" + jobId + "/events?after=-1&limit=1000")
@@ -194,7 +195,7 @@ class SearchModeSearchAgentTest {
         var jobId = createJobWithSpec(jobSpec, "search-test-scan-model");
 
         // Wait for job to complete
-        awaitJobCompletion(jobId, Duration.ofSeconds(30));
+        awaitJobCompletion(baseUrl, jobId, authToken, Duration.ofSeconds(30));
 
         // Poll events to ensure search produced LLM tokens/notifications and no code edits
         var eventsUrl = URI.create(baseUrl + "/v1/jobs/" + jobId + "/events?after=-1&limit=1000")
@@ -276,7 +277,7 @@ class SearchModeSearchAgentTest {
         var jobId = createJobWithSpec(jobSpec, "search-test-ignore-code-model");
 
         // Wait for job to complete
-        awaitJobCompletion(jobId, Duration.ofSeconds(30));
+        awaitJobCompletion(baseUrl, jobId, authToken, Duration.ofSeconds(30));
 
         // Verify job status
         var statusUrl = URI.create(baseUrl + "/v1/jobs/" + jobId).toURL();
@@ -297,29 +298,6 @@ class SearchModeSearchAgentTest {
     // ============================================================================
     // Helpers
     // ============================================================================
-
-    private void awaitJobCompletion(String jobId, Duration timeout) throws Exception {
-        long deadline = System.currentTimeMillis() + timeout.toMillis();
-        while (System.currentTimeMillis() < deadline) {
-            var statusUrl = URI.create(baseUrl + "/v1/jobs/" + jobId).toURL();
-            var conn = (HttpURLConnection) statusUrl.openConnection();
-            conn.setRequestMethod("GET");
-            conn.setRequestProperty("Authorization", "Bearer " + authToken);
-
-            try (InputStream is = conn.getInputStream()) {
-                var response = new String(is.readAllBytes(), StandardCharsets.UTF_8);
-                if (response.contains("\"state\":\"COMPLETED\"")
-                        || response.contains("\"state\":\"FAILED\"")
-                        || response.contains("\"state\":\"CANCELLED\"")) {
-                    return;
-                }
-            } finally {
-                conn.disconnect();
-            }
-            Thread.sleep(100);
-        }
-        throw new AssertionError("Job did not complete within timeout: " + timeout);
-    }
 
     private byte[] createEmptyZip() throws IOException {
         var out = new java.io.ByteArrayOutputStream();
